Higher_Lower_Same
"Higher or Lower" is a simple and engaging Command Line Interface (CLI) card game built with Python. The objective is to guess whether the next card drawn from the deck will be higher, lower, or the same as the current card. The game continues until the deck runs out or the player decides to quit.